configure: enable LTTng with --enable-lttng-ust option
LTTng is a tool for low overhead tracing of events. It is already
employed by QEMU and friendly with tracing mechanism of linux kernel,
so it is very useful for performance analysis and debugging of
sheepdog.

Below is a brief example of lttng tracing:

$ lttng create sheepdog
Session sheepdog created.
Traces will be written in /home/mitake/lttng-traces/sheepdog-20150109-161938
$ lttng enable-event -a -u -c sheepdog
All UST events are enabled in channel sheepdog
$ lttng start
Tracing started for session sheepdog
 # do something you want to analyze
$ lttng stop

Then, you can view or analyze collected events in your favorite way.
Example of babeltrace (just viewing):
$ babeltrace lttng-traces/sheepdog-20150109-161938

This patch is just a preparation. It adds a new option
--enable-lttng-ust for activating LTTng ready binaries of sheepdog.

Cc: Gu Ping <guping610@qq.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>