sheep, dog: fast deep copy for snapshot
dog vdi snapshot <vdi> --no-share has a bottleneck: the dog
process. This patch adds a new option --fast-deep-copy to dog vdi
snapshot, which avoid the bottleneck.

If the option is passed to the dog command, actual copying of data
objects are done by sheep processes. So the dog process isn't a
bottleneck in this case.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>