dog: type cast miss at vdi_show_progress
The result of the multiplication of uint32 and uint64 are rounded to uint32
So, Progress of the command is displayed sometimes back to 0%

This is only a display issue, command processing can be completed.

necessary in show_progress of calculation, change type of idx to uint64_t.
However, idx in the current Sheepdog is within 32bit number,
so upper limit check idx, using the max_idx can help us avoid the side effect.

Signed-off-by: Masaki Saeki <saeki.masaki@po.ntts.co.jp>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>