dog: add a new option for reducing identical snapshots
Current "dog vdi snapshot" command creates a new snapshot
unconditionally, even if a working VDI doesn't have its own
objects. In such a case, the created snapshot is redundant because
same VDI is already existing.

This patch adds a new option -R to the dog command for reducing
the identical snapshots.

Motivations:
1. why we need it in dog command instead of combination of parsing
   "dog vdi list" and existing "dog vdi snapshot"?
"dog vdi list" tends to be slow in a cluster which has thousands of
VDIs because it issues bunch of read requests for inode headers. It
should be avoided as much as possible.
2. why identical snapshots harmful?
Creating identical snapshots means increasing a number of inode
objects (especially in EC, 16:4 VDI creates 20 inodes). It harms
performance of recovery process. And exhaustion of VID space will come
early.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>