zookeeper: show real session timeout in zk_init()
ZK server has its own session timeout range, so client's timeout value passed
to server might be ignored and return the value it thinks proper.

From the documentation, the server range is

[tickTime * 2, tickTime * 20]

e.g, if tickTime is 2s, then the range will be [4s, 40s]. If we pass 1s as
timeout to zk driver, we might be 4s returned.

This patch also change the zk server informatin printed as sd_info.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>