zookeeper: multi-sheep cluster in one zookeeper support
An optional "chroot" suffix can be appended to the connection
string(or hosts list). This will run the client commands while
interpreting all paths relative to this root.
Different sheep clusters have their own root znode,
so one zk can support muilt-sheep cluster.
If it is not set, the default root is /sheepdog,
which is compatible with old versions.

Signed-off-by: Yu Yang <yuyang@cmss.chinamobile.com>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>