shared lib: add the low level request handling framework
The core idea is the same as sbd's framework.

         User request
             |
             | aio control block (struct aiocb)
             V
        +----+----+
        |    |    |
       r1   r2    r3
        |    |    |
        V    V    V
       obj  obj  obj

User request is spit into several sheep requests which are performed on the
individual sheepdog object. All the sheep requests are queued and sent in a
async mode concurently. When all the sheep requests are done, aiocb will try
to user of the completion of the user request.

For future possible other administritive opcodes, we provide a sync function,
which can run the opcode and get the response in a sync mode.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>