Various init script fixups
Source init-functions on deb-based distros to provide killproc.

Use SHEEPDOGD variable in init script

SHEEPDOGD is defined but is not used when starting sheep. This change
allows system default configs to override SHEEPDOGD if desired.

Use SHEEPDOG_PATH in init script

The debian defaults file has a comment that this environment variable
is used by the init script to set the sheepdog storage dir, however
the code in the init script does not support this claim. This change
makes the needed changes that allow this to work.