dog: upgrade command for epoch log file
This patch adds a new upgrade command for epoch log file. It can be
used as below:
$ dog upgrade epoch-convert <original epoch file path> <new file path>

Example of output:
number of vnodes of each nodes:
        IPv4 ip:127.0.0.1 port:7000 == 64
        IPv4 ip:127.0.0.1 port:7001 == 64
        IPv4 ip:127.0.0.1 port:7002 == 64
please supply the above numbers to sheeps with -V option

As the output says, these numbers must be supplied to new sheep
processes with -V option for enabling fixedvnodes storategy. In other
words, the upgraded cluster cannot use autovnode strategy.

Cc: Masahiro Tsuji <tuji@atworks.co.jp>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>