sheep: wildcard recovery strategy
This patch adds a new type of recovery strategy, wildcard. Current
sheepdog doesn't have a method for migrating objects when it changes
hash function for object placement. The command added by the previous
patch is a solution for the problem, but it requires manual copy
spanning server.

As an alternative and more easy way to solve the problem, this patch
adds a new option -W to sheep command. When the option is passed,
sheep searches all objects from all nodes. Therefore, a cluster can be
upgraded even if object placement strategy is changed (e.g. from 0.7.x
to 0.9.y).

The feature will be useful for cases like switching normal mode to
diskvnodes in the future.

TODO: erasure coded VDIs are not supported yet.

Cc: Masahiro Tsuji <tuji@atworks.co.jp>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>

Conflicts:
	sheep/recovery.c
	sheep/sheep.c