sheep: avoid panic in COW when sheep fails to allocate memory
Current code for COW requests can call panic() because it uses
xzalloc() and xcalloc() for processing the requests. This commit lets
the code avoid use the functions and return SD_RES_NO_MEM because COW
is a very common and frequent request. It is better to return error
code to VMs than killing sheep.

Even if the memory allocations fail, it just causes object leaks. This
behavior is not changed from the existing strategy which calls
panic().

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>