add -d option for dog cluster info
Dog cluster info with option -d just outputs the changed nodes
between the two adjacent epoches. It's useful in a cluster with
a large amount of nodes. And it would be a way easier to create
a nagios/icinga script that advertises you when something changes
in the cluster and what has changed.

Signed-off-by: Meng Lingkun <menglingkun@cmss.chinamobile.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>