SystemD doesn't listen to ulimit
SystemD doesn't listen to anyone's ulimit methods. It forces the daemon and it's children to 1024. There is a complex way of allocating more resources to a 'slice' using additional files but I've found that "LimitNOFILE" works as well after a reboot of the entire system, not just a daemon-reload or service restart.