lib/util: Fix xrealloc to check out of memory error correctly
The present xrealloc panics as "Out of memory" if realloc returns NULL.
However, man page of realloc[1] says that "If size was equal to 0, either
NULL or a pointer suitable to be passed to free() is returned." So xrealloc
can panic wrongly when size is equal to 0 and realloc returns NULL.

This commit fixes the issue described above. The man page also says that
"The UNIX 98 standard requires [..] realloc() to set errno to ENOMEM upon
failure." So I evaluate errno instead of returned value to check whether
out of memory really occurred or not.

[1] http://linux.die.net/man/3/realloc

Signed-off-by: Takashi Menjo <menjo.takashi@lab.ntt.co.jp>