Revert "sheep: update ledger objects in an asynchronous manner"
As mentioned in #362, there is a risk of deadlock in the stable-0.9
branch. We can mitigate (or may remove) the risk by increasing the
maximum number of threads for workqueues enough. However, it is hard
to forecast the behavior of parallel execution of (1) and (2) below,
making it difficult to estimate the number of threads:

  (1) gateway_write_object (run in a "gateway" thread) to update the
      data_vdi_id field of an inode, causing update of reference count

  (2) async_update_obj_refcnt_work (run in an "io" thread) to update
      reference count, following (1)

This commit lets the two above run in serial in a "gateway" thread.
Note that the (2) is now update_obj_refcnt.

This reverts commit 796c88df19e32a6a83a6ba54f2b17b03e166b3d2.

Signed-off-by: Takashi Menjo <menjo.takashi@lab.ntt.co.jp>

Conflicts:
	sheep/gateway.c