tests: fix functional 114 to take into account the block of SELinux.
If SELinux is not disabled (enforcing or permissive), functional test
No.114 will fail.
This is because, a single block for SELinux's attribute is added to
the '%b * %B' of stat command output, so the output result is different
from the SELinux disabled environment.

This change provides, it will take into account the block of SELinux.

Signed-off-by: Kazuhisa Hara <khara@sios.com>